gossip.TxPosition --> app.TxPosition
